Moved all functions except display type functions to seperate namespaces
