+++ Proj/Project_1/P1_The_War_Game/Constants.h
@@ -1,0 +1,24 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+/* 
+ * File:   Constants.h
+ * Author: Kevin Maltz
+ *
+ * Created on April 21, 2018, 6:06 PM
+ */
+
+#ifndef CONSTANTS_H
+#define CONSTANTS_H
+
+const short ROW_MX = 13;//Maximum number of rows on the board
+const short COL_MX = 5; //Maximum number of columns on the board
+const int N_PCS = 25;   //Number of pieces per player
+const int RM_CL = 13;   //Must be = to the c of Location.dsply[r][c]
+const int RM_RW = 5;    //Must be = to the r of Location.dsply[r][c]
+
+#endif /* CONSTANTS_H */
+
