+++ Proj/Project_1/P1_The_War_Game/Location.h
@@ -14,13 +14,14 @@
 #ifndef LOCATION_H
 #define LOCATION_H
 
+#include "Constants.h"
 #include "Unit.h"
 struct Location{
     char type;      //What type of position is it. Camp/Battlefield/Frontline/Headquarters/Mountain
     bool isOcp;     //If location occupied, set true
     bool isRR;      //Set true if location is on railroad line
-    char dsply1[5][13]; //Display for player 1
-    char dsply2[5][13]; //Display for player 2
+    char dsply1[RM_RW][RM_CL]; //Display for player 1
+    char dsply2[RM_RW][RM_CL]; //Display for player 2
     Unit *occUnit;  //Pointer to unit occupying the location.
 };
