+++ Proj/Project_1/P1_The_War_Game/main.cpp
@@ -13,8 +13,11 @@
 using namespace std;
 
 //User Libraries
+#include "Constants.h"
 #include "Unit.h"
 #include "Location.h"
+#include "setup.h"
+#include "play.h"
 //Structures
 //struct Unit{
 //    unsigned short priority;  //The value representing its position on the attack hierarchy, higher is better
@@ -32,36 +35,36 @@ using namespace std;
 //};
 
 //Global Constants - Math, Science, Conversions, 2D Array Sizes
-const short ROW_MX = 13;//Maximum number of rows on the board
-const short COL_MX = 5; //Maximum number of columns on the board
-const int N_PCS = 25;   //Number of pieces per player
-const int RM_CL = 13;   //Must be = to the c of Location.dsply[r][c]
-const int RM_RW = 5;    //Must be = to the r of Location.dsply[r][c]
+//const short ROW_MX = 13;//Maximum number of rows on the board
+//const short COL_MX = 5; //Maximum number of columns on the board
+//const int N_PCS = 25;   //Number of pieces per player
+//const int RM_CL = 13;   //Must be = to the c of Location.dsply[r][c]
+//const int RM_RW = 5;    //Must be = to the r of Location.dsply[r][c]
 
 //Function Prototypes
-Location **initBrd(fstream &);  //Initializes the board
-void initPcs(fstream &, Unit [], Unit []);  //Initialize data for all pieces(Unit structures)
-int stPlyrs(string &, string &, bool &);    //Sets the players and their names, returns number of human players
-void setPcs(Location **, Unit [], string);
-void ocpy(Location*, Unit*);    //Occupy a location with the passed Unit, modify displays of location to reflect
-void unOcpy(Location*);         //Remove Unit from passed location, remove unit from displays as well
+//Location **initBrd(fstream &);  //Initializes the board
+//void initPcs(fstream &, Unit [], Unit []);  //Initialize data for all pieces(Unit structures)
+//int stPlyrs(string &, string &, bool &);    //Sets the players and their names, returns number of human players
+//void setPcs(Location **, Unit [], string);
+//void ocpy(Location*, Unit*);    //Occupy a location with the passed Unit, modify displays of location to reflect
+//void unOcpy(Location*);         //Remove Unit from passed location, remove unit from displays as well
 void dspBrd(Location **, int);  //Displays the board
 void dspHP(int);                //displays the horizontal paths on the board
 void dspVP1(int);               //displays the vertical paths on the board for player 1
 void dspVP2(int);               //displays the vertical paths on the board for player 2
-bool mveVld(Location **, int, int, int, int, Unit *);
-bool isRRopn(Location **, int, int, int, char);
-void combat(Unit *, Location &); //Determines combat results
-void move(Location **, int pTurn);    //Move a piece from one location to another
-string frmtCse(string);         //Formats string to proper noun capitalization
-int fndMtch(Unit[], string);                  //Get index of matching unit in array, if exists
-void whoNtSt(Unit[]);           //Lists all pieces not yet in play for a single player
-void ptBrdLoc(Location **);     //Test if binary file properly read to board structure array
-void ptPlyrs(Unit [], Unit []); //Test if player pieces read in successfully
+//bool mveVld(Location **, int, int, int, int, Unit *);
+//bool isRRopn(Location **, int, int, int, char);
+//void combat(Unit *, Location &); //Determines combat results
+//void move(Location **, int pTurn);    //Move a piece from one location to another
+//string frmtCse(string);         //Formats string to proper noun capitalization
+//int fndMtch(Unit[], string);                  //Get index of matching unit in array, if exists
+//void whoNtSt(Unit[]);           //Lists all pieces not yet in play for a single player
+//void ptBrdLoc(Location **);     //Test if binary file properly read to board structure array
+//void ptPlyrs(Unit [], Unit []); //Test if player pieces read in successfully
 
 //Inlined functions
-inline bool XOR(int lh, int rh){return (!(lh && rh)) && (lh || rh);}
-inline bool XOR(bool lh, bool rh){return (!(lh && rh)) && (lh || rh);}
+//inline bool XOR(int lh, int rh){return (!(lh && rh)) && (lh || rh);}
+//inline bool XOR(bool lh, bool rh){return (!(lh && rh)) && (lh || rh);}
 
 //Execution begins here
 int main(int argc, char** argv) 
@@ -72,17 +75,17 @@ int main(int argc, char** argv)
     bool ai = false;        //Playing against AI?
     unsigned int turn = 0;  //The current turn number. Increments +1 every turn
     
-    fstream setup("setup.dat", ios::in | ios::binary);
+    fstream stpInp("setup.dat", ios::in | ios::binary);
     Location **board;   //13x5 game board
     Unit p1Pcs[N_PCS];  //Holds all of player 1's pieces
     Unit p2Pcs[N_PCS];  //Holds all of player 2's pieces
     
-    if(setup.is_open() && setup.good()){
+    if(stpInp.is_open() && stpInp.good()){
         //Read in data for every location on the board
-        board = initBrd(setup);
+        board = setup::initBrd(stpInp);
         //Read in data for each piece
-        initPcs(setup, p1Pcs, p2Pcs);
-        setup.close();
+        setup::initPcs(stpInp, p1Pcs, p2Pcs);
+        stpInp.close();
     }
     else{   //Return exit failure if setup file fails to open.
         cout << "ERROR: unable to open file Setut.dat, exiting program\n";
@@ -90,14 +93,13 @@ int main(int argc, char** argv)
     }
     
     //Set player(s) and return number of human players
-    nPlyr = stPlyrs(plyr1, plyr2, ai);
+    nPlyr = setup::stPlyrs(plyr1, plyr2, ai);
     //Display the board, right now with no pieces.
-    dspBrd(board,1);
+    dspBrd(board,2);
     //player 1 piece setup
-    setPcs(board, p1Pcs, plyr1);
-            
+    setup::setPcs(board, p1Pcs, plyr1);
     //player 2 piece setup
-    setPcs(board, p2Pcs, plyr2);
+    setup::setPcs(board, p2Pcs, plyr2);
 //TODO - /*FOR AI IMPLEMENTATION*/
 //    if(ai)
 //        aiStPcs(board, p2Pcs);
@@ -117,199 +119,85 @@ int main(int argc, char** argv)
     return EXIT_SUCCESS;
 }
 
-Location **initBrd(fstream &setup){
-    Location **board;
-    board = new Location*[ROW_MX];
-    for(int r=0; r < ROW_MX; r++){
-        board[r] = new Location[COL_MX];
-        for(int c=0; c < COL_MX; c++){   
-            //Read in binary data to a location structure
-            setup.read(reinterpret_cast<char *>(board[r]+c), sizeof(board[r][c]));
-            //Ensure all Unit* set to NULL. Ensures no bad memory read in from file
-            board[r][c].occUnit = NULL;
-        }
-    }
-    return board;
-}
-void initPcs(fstream &setup, Unit p1[], Unit p2[]){
-    int nmLngth;    //Length of string to be read in from binary, minus the null
-    char* buf;      //Buffer for reading in a string from the binary file
-    for(int i=0; i < N_PCS; i++){
-            setup.read(reinterpret_cast<char*>(&nmLngth), sizeof(nmLngth));
-            setup.read(reinterpret_cast<char*>(&(p1[i].priority)), sizeof(p1[i].priority));
-            setup.read(reinterpret_cast<char*>(&(p1[i].inPlay)), sizeof(p1[i].inPlay));
-            buf = new char[nmLngth + 1];
-            setup.read(buf,nmLngth);
-            buf[nmLngth] = '\0';
-            p1[i].name = buf;
-            
-            //Copy over Unit structure from p1 to p2
-            p2[i].priority = p1[i].priority;
-            p2[i].inPlay = p1[i].inPlay;
-            p2[i].name = p1[i].name;
-            
-            //Set ownership of pieces
-            p1[i].plyrID = 1;
-            p2[i].plyrID = 2;
-            
-            //Release memory before next iteration to prevent memory leaks
-            delete [] buf;
-        }
-}
-int stPlyrs(string &plyr1, string &plyr2, bool &plyNPC){
-    int nPlyr = 0;  //Number of human players. Sentinel value of 0
-    
-    cout << "2 player game or 1 player vs AI available\n";
-    cout << "Enter number of players(2 max): ";
-    cin >> nPlyr;
-    cin.ignore(1000, '\n');
-    while(nPlyr != 1 && nPlyr != 2){
-        cout << "Invalid number of player. Re-enter(max 2): ";
-        cin >> nPlyr;
-        cin.ignore(1000, '\n');
-    }
-    if(nPlyr == 1){
-        cout << "Enter player name: ";
-        getline(cin, plyr1);
-        plyr2 = "NPC";
-        plyNPC = true;
-    }
-    if(nPlyr == 2){
-        cout << "Enter player 1's name: ";
-        getline(cin, plyr1);
-        cout << "Enter player 2's name: ";
-        getline(cin, plyr2);
-    }
-    return nPlyr;
-}
-void setPcs(Location **board, Unit pcs[], string pNme){
-    string slctn;   //piece selection entered by player
-    int mtchInd;    //Index of Unit.name in pxPcs that matches slctn. -1 sentinel value
-    bool allSet;    //Stays true if all pieces have been set on the board
-    int slctR, slctC;   //selected row and column of location to place a piece
-    char rpPc;      //Choice to replace a piece if location is already occupied
-    bool repeat;    //General boolean to use if loops need repeating
-    const int R_MX_P1 = 5;  //Maximum row index player 1 can set pieces at
-    const int R_MN_P2 = 7;  //Minimum row index player 2 can set pieces at
-    int id = pcs[0].plyrID; //Which player is setting their pieces
-    
-    cout << endl << endl;
-    cout << pNme << " place your pieces on the board.\n";
-    //Begin setup
-    do{
-        allSet = true;
-        mtchInd = -1;
-        cout << "Unplaced pieces:\n";
-        //Display all pieces not yet in play/set on board
-        whoNtSt(pcs);
-        //Get and validate piece selection by player
-        do{
-            repeat = false;
-            //Prompt for choice of piece
-            cout << pNme << " select a piece by typing in the name: ";
-            getline(cin, slctn);
-            //Convert string to proper case formatting
-            slctn = frmtCse(slctn);
-            //Check if piece exists and not in play. If valid store index of selected unit
-            mtchInd = fndMtch(pcs, slctn);
-            //If no valid match found notify player and restart piece selection
-            if(mtchInd == -1){
-                cout << "Invalid selection\n";
-                repeat = true;
-            }
-        }while(repeat);//FIND ME        
-        //Setup pieces on board
-        do{
-            repeat = false;;
-            do{ //Get and validate location selection. Repeats till location is valid
-                repeat = false;
-                slctR = slctC = -1;
-                cout << "Enter row,column( #,# ) to select a location to place the " 
-                     << pcs[mtchInd].name << ": ";
-                //TODO - change this to use a stringstream to catch any error in the way user inputs numbers
-                cin >> slctR; cin.ignore(1000,','); 
-                cin >> slctC; cin.ignore(1000,'\n');
-                //Valid row of locations depend on the player ID
-                if(id == 1 && (slctR < 0 || slctR > R_MX_P1)){
-                    repeat = true;
-                    cout << "Invalid location\n";
-                }
-                else if(id == 2 && (slctR < R_MN_P2 || slctR >= ROW_MX)){
-                    repeat = true;
-                    cout << "Invalid location\n";
-                }
-                else if(slctC < 0 || slctC >= COL_MX){
-                    repeat = true;
-                    cout << "Invalid location\n";
-                }
-            }while(repeat);            
-            //Check if chosen location is already occupied, if yes decide what to do
-            if(board[slctR][slctC].occUnit != NULL){
-                cout << board[slctR][slctC].occUnit->name
-                     << " is already at that location.\nDo you wish to replace it with "
-                     << pcs[mtchInd].name << "? Enter Y or N: ";
-                cin >> rpPc; cin.ignore(1000,'\n');
-                
-                //Repeat location selection process if no, replace piece otherwise
-                if(rpPc == 'n' || rpPc == 'N')
-                    repeat = true;
-                else if(rpPc == 'y' || rpPc == 'Y')
-                    unOcpy(&board[slctR][slctC]);
-                else{
-                    cout << "Invalid entry. Replacing piece anyways.\n";
-                    unOcpy(&board[slctR][slctC]);
-                }
-            }
-        }while(repeat);
-        ocpy(&board[slctR][slctC], &pcs[mtchInd]);        
-        //If any pieces are not yet set in play, allSet=false and loop setup
-        for(int i=0; i < N_PCS && allSet; i++)
-            if(pcs[i].inPlay == false){
-                allSet = false;
-                break;  //Stop looking as soon as any unset piece is found
-            }
-    }while(!allSet);
-}
-void ocpy(Location* loc, Unit* unit){
-    //Occupy location, place unit at location, set unit as in play
-    loc->isOcp = true;
-    loc->occUnit = unit;
-    unit->inPlay = true;
-    //Depending on owner of unit, configure board display at location for each player
-    if(unit->plyrID == 1){
-        if(unit->name == "Field Marshal")
-            for(int c=0; c < 7; c++)
-                loc->dsply1[2][2+c] = unit->name[6+c];
-        else
-            for(int c=0; c < unit->name.length(); c++)
-                loc->dsply1[2][2+c] = unit->name[c];
-        for(int c=2; c < RM_CL-2; c++)
-            loc->dsply2[2][c] = '?';
-            
-    }
-    else{
-        if(unit->name == "Field Marshal")
-            for(int c=0; c < 7; c++)
-                loc->dsply2[2][2+c] = unit->name[6+c];
-        else
-            for(int c=0; c < unit->name.length(); c++)
-                loc->dsply2[2][2+c] = unit->name[c];
-        for(int c=2; c < RM_CL-2; c++)
-            loc->dsply1[2][c] = '?';
-    }
-}
-void unOcpy(Location* loc){
-    //Set unit as out of play, remove unit from location, un-occupy location
-    loc->occUnit->inPlay = false;
-    loc->occUnit = NULL;
-    loc->isOcp = false;
-    //Clear board display of location for both players
-    for(int r=1; r < RM_RW-1; r++)
-        for(int c=1; c < RM_CL-1; c++){
-            loc->dsply1[r][c] = ' ';
-            loc->dsply2[r][c] = ' ';
-        }
-}
+//Location **initBrd(fstream &setup){
+//    Location **board;
+//    board = new Location*[ROW_MX];
+//    for(int r=0; r < ROW_MX; r++){
+//        board[r] = new Location[COL_MX];
+//        for(int c=0; c < COL_MX; c++){   
+//            //Read in binary data to a location structure
+//            setup.read(reinterpret_cast<char *>(board[r]+c), sizeof(board[r][c]));
+//            //Ensure all Unit* set to NULL. Ensures no bad memory read in from file
+//            board[r][c].occUnit = NULL;
+//        }
+//    }
+//    return board;
+//}
+//void initPcs(fstream &setup, Unit p1[], Unit p2[]){
+//    int nmLngth;    //Length of string to be read in from binary, minus the null
+//    char* buf;      //Buffer for reading in a string from the binary file
+//    for(int i=0; i < N_PCS; i++){
+//            setup.read(reinterpret_cast<char*>(&nmLngth), sizeof(nmLngth));
+//            setup.read(reinterpret_cast<char*>(&(p1[i].priority)), sizeof(p1[i].priority));
+//            setup.read(reinterpret_cast<char*>(&(p1[i].inPlay)), sizeof(p1[i].inPlay));
+//            buf = new char[nmLngth + 1];
+//            setup.read(buf,nmLngth);
+//            buf[nmLngth] = '\0';
+//            p1[i].name = buf;
+//            
+//            //Copy over Unit structure from p1 to p2
+//            p2[i].priority = p1[i].priority;
+//            p2[i].inPlay = p1[i].inPlay;
+//            p2[i].name = p1[i].name;
+//            
+//            //Set ownership of pieces
+//            p1[i].plyrID = 1;
+//            p2[i].plyrID = 2;
+//            
+//            //Release memory before next iteration to prevent memory leaks
+//            delete [] buf;
+//        }
+//}
+//void ocpy(Location* loc, Unit* unit){
+//    //Occupy location, place unit at location, set unit as in play
+//    loc->isOcp = true;
+//    loc->occUnit = unit;
+//    unit->inPlay = true;
+//    //Depending on owner of unit, configure board display at location for each player
+//    if(unit->plyrID == 1){
+//        if(unit->name == "Field Marshal")
+//            for(int c=0; c < 7; c++)
+//                loc->dsply1[2][2+c] = unit->name[6+c];
+//        else
+//            for(int c=0; c < unit->name.length(); c++)
+//                loc->dsply1[2][2+c] = unit->name[c];
+//        for(int c=2; c < RM_CL-2; c++)
+//            loc->dsply2[2][c] = '?';
+//            
+//    }
+//    else{
+//        if(unit->name == "Field Marshal")
+//            for(int c=0; c < 7; c++)
+//                loc->dsply2[2][2+c] = unit->name[6+c];
+//        else
+//            for(int c=0; c < unit->name.length(); c++)
+//                loc->dsply2[2][2+c] = unit->name[c];
+//        for(int c=2; c < RM_CL-2; c++)
+//            loc->dsply1[2][c] = '?';
+//    }
+//}
+//void unOcpy(Location* loc){
+//    //Set unit as out of play, remove unit from location, un-occupy location
+//    loc->occUnit->inPlay = false;
+//    loc->occUnit = NULL;
+//    loc->isOcp = false;
+//    //Clear board display of location for both players
+//    for(int r=1; r < RM_RW-1; r++)
+//        for(int c=1; c < RM_CL-1; c++){
+//            loc->dsply1[r][c] = ' ';
+//            loc->dsply2[r][c] = ' ';
+//        }
+//}
 void dspBrd(Location **board, int pID){
     //Always display the current players' board half on bottom
     if(pID == 1){
@@ -519,200 +407,200 @@ void dspVP2(int brdRow){
 
     }
 }
-void combat(Unit *attckr, Location &loc){
-    if(attckr->priority == 10){
-        //Bomb blows up, so do you!
-        unOcpy(&loc);
-        attckr->inPlay = false;
-    }
-    else if(attckr->name == "Engineers" && loc.occUnit->name == "Landmines"){
-        //Remove Landmines
-        unOcpy(&loc);
-        //Place Engineers at the location
-        ocpy(&loc, attckr);
-    }
-    else{
-        if(attckr->priority > loc.occUnit->priority){
-            //Defender dies, invasion succeeds
-            unOcpy(&loc);
-            ocpy(&loc, attckr);
-        }
-        else if(attckr->priority < loc.occUnit->priority){
-            //Attacker dies, invasion fails
-            attckr->inPlay = false;
-        }
-        else if(attckr->priority == loc.occUnit->priority){
-            //Mutual destruction, location is now empty
-            unOcpy(&loc);
-            attckr->inPlay = false;
-        }
-    }
-}
-bool mveVld(Location **board, int strtR, int strtC, int desR, int desC, int id){
-    bool vld = true;        //True by default, set false if invalid move found
-    int dspR = strtR - desR;//Absolute Row displacement of unit
-    int dspC = strtC - desC;//Absolute Column displacement of unit
-    //Take the absolute value of the displacement
-    if(dspR < 0)
-        dspR *= -1;
-    if(dspC < 0)
-        dspC *= -1;
-    //Check if both starting location and destination are on the board
-    if(strtR<0 || strtC<0 || desR<0 || desC<0)
-        vld = false;
-    else if(strtR >= ROW_MX || strtC >= COL_MX || desR >= ROW_MX || desC >= COL_MX)
-        vld = false;
-    //Check if Frontline or Mountains were selected as destination
-    else if(board[desR][desC].type == 'M' || board[desR][desC].type == 'F')
-        vld = false;
-    //Check if location is occupied that it is not players own unit
-    else if(board[desR][desC].isOcp && board[desR][desC].occUnit->plyrID == id)
-        vld = false;
-    //If move is <1,1> then make sure a Camp location is invloved
-    else if(dspR == 1 && dspC == 1){
-        if(board[strtR][strtC].type != 'C' && board[desR][desC].type != 'C')
-            vld = false;
-    }
-    else if(dspR > 1 && dspC > 1)    //TODO - place the engineer exception in here later
-        vld = false;
-    else if(XOR(dspR > 1, dspC > 1) && board[strtR][strtC].isRR){
-        if(!board[desR][desC].isRR)
-            vld = false;
-        else if(dspR > 1)
-            vld = isRRopn(board, strtR, strtC, desR, 'r');
-        else
-            vld = isRRopn(board, strtR, strtC, desC, 'c');
-    }
-    return vld;
-}
-bool isRRopn(Location **board, int initR, int initC, int dest, char drctn)
-{
-    bool rrOpn = true;
-    if(drctn == 'c'){
-        if(dest - initC > 0){
-            for(int c=initC+1; c < dest && rrOpn; c++)
-                if(board[initR][c].isOcp || !board[initR][c].isRR)
-                    rrOpn = false;
-        }
-        else if(dest - initC < 0)
-            for(int c=initC-1; c > dest && rrOpn; c--)
-                if(board[initR][c].isOcp || !board[initR][c].isRR)
-                    rrOpn = false;
-    }
-    else if(drctn == 'r'){
-        if(dest - initR > 0){
-            for(int r=initR+1; r < dest && rrOpn; r++)
-                if(board[r][initC].isOcp || !board[r][initC].isRR)
-                    rrOpn = false;
-        }
-        else if(dest - initR < 0)
-            for(int r=initR-1; r > dest && rrOpn; r--)
-                if(board[r][initC].isOcp || !board[r][initC].isRR)
-                    rrOpn = false;
-    }
-    return rrOpn;
-}
-void move(Location **board, int pTurn){
-    bool vldMve;
-    int strtR, strtC;   //(R,C) of starting position
-    int desR, desC;     //(R,C) of destination position
-    Unit *unit;
-    //Loop until valid move is entered
-    do{
-        //Prompt player for movement
-        cout << "Enter the location of the piece you want to move. Use R,C format: ";
-        cin >> strtR; cin.ignore(1000,','); 
-        cin >> strtC; cin.ignore(1000,'\n');
-        cout << "Enter your destination in R,C format: ";
-        cin >> desR; cin.ignore(1000,','); 
-        cin >> desC; cin.ignore(1000,'\n');
-        //Validate movement
-        vldMve = mveVld(board, strtR, strtC, desR, desC, pTurn);
-        if(!vldMve)
-            cout << "Invalid Move\n";
-        if(board[strtR][strtC].isOcp){
-            unit = board[strtR][strtC].occUnit;
-            if(unit->plyrID != pTurn){
-            cout << "That is not your piece!\n";
-            vldMve = false;
-            }
-            else if(unit->name == "Flag" || unit->name == "Landmines"){
-                cout << unit->name << " cannot ever be moved.\n";
-                vldMve = false;
-            }
-        }
-        else{
-            cout << "No piece available to move from that location.\n";
-            vldMve = false;
-        }
-    }while(!vldMve);
-    //Now make the move, including combat
-    if(board[desR][desC].isOcp)
-        combat(unit, board[desR][desC]);
-    else{
-        unOcpy(&board[strtR][strtC]);
-        ocpy(&board[desR][desC], unit);
-    } 
-}
-string frmtCse(string s){
-    s[0] = toupper(s[0]);
-    for(int i=1; i<s.length(); i++)
-        s[i] = tolower(s[i]);
-
-    //Proper capitalization for Field Marshal
-    if(s.length() == 13)
-        s[6] = toupper(s[6]);
-    
-    return s;
-}
-int fndMtch(Unit pcs[], string s){
-    for(int i=0; i < N_PCS; i++){
-        if(pcs[i].inPlay == false)
-            if(pcs[i].name == s)
-                return i; //return value as soon as a match is found
-    }
-    return -1;  //Flag for no match found
-}
-void whoNtSt(Unit pcs[]){
-    int perLne = 0;
-    for(int i=0; i < N_PCS; i++){
-        if(pcs[i].inPlay == false){
-            cout << setw(14) << pcs[i].name;
-            perLne++;
-            if(0 == (perLne % 3))
-                cout << endl;
-        }
-    }
-    cout << endl;
-}
-void ptBrdLoc(Location **brd){
-    for(int r=0; r < ROW_MX; r++){
-        for(int c=0; c < COL_MX; c++)
-            cout << brd[r][c].type << " ";
-        cout << endl;
-        for(int c=0; c < COL_MX; c++)
-            cout << brd[r][c].isOcp << " ";
-        cout << endl;
-        for(int c=0; c < COL_MX; c++)
-            cout << brd[r][c].isRR << " ";
-        cout << endl;
-        for(int c=0; c < COL_MX; c++)
-            cout << brd[r][c].occUnit << " ";
-        cout << endl;
-        for(int i=0; i < RM_CL; i++){
-            for(int c=0; c < COL_MX; c++)
-                for(int j=0; j < RM_RW; j++){
-                    cout << brd[r][c].dsply1[i][j];
-                    if(j==12)
-                        cout << "     ";
-                }
-            cout << endl;
-        }
-    }
-}
-void ptPlyrs(Unit p1[], Unit p2[]){
-    for(int i=0; i < N_PCS; i++){
-        cout << "P1: " << p1[i].priority << " " << p1[i].inPlay << " " << p1[i].name << endl;
-        cout << "P2: " << p2[i].priority << " " << p2[i].inPlay << " " << p2[i].name << endl;
-    }
-}
+//void combat(Unit *attckr, Location &loc){
+//    if(attckr->priority == 10){
+//        //Bomb blows up, so do you!
+//        unOcpy(&loc);
+//        attckr->inPlay = false;
+//    }
+//    else if(attckr->name == "Engineers" && loc.occUnit->name == "Landmines"){
+//        //Remove Landmines
+//        unOcpy(&loc);
+//        //Place Engineers at the location
+//        ocpy(&loc, attckr);
+//    }
+//    else{
+//        if(attckr->priority > loc.occUnit->priority){
+//            //Defender dies, invasion succeeds
+//            unOcpy(&loc);
+//            ocpy(&loc, attckr);
+//        }
+//        else if(attckr->priority < loc.occUnit->priority){
+//            //Attacker dies, invasion fails
+//            attckr->inPlay = false;
+//        }
+//        else if(attckr->priority == loc.occUnit->priority){
+//            //Mutual destruction, location is now empty
+//            unOcpy(&loc);
+//            attckr->inPlay = false;
+//        }
+//    }
+//}
+//bool mveVld(Location **board, int strtR, int strtC, int desR, int desC, int id){
+//    bool vld = true;        //True by default, set false if invalid move found
+//    int dspR = strtR - desR;//Absolute Row displacement of unit
+//    int dspC = strtC - desC;//Absolute Column displacement of unit
+//    //Take the absolute value of the displacement
+//    if(dspR < 0)
+//        dspR *= -1;
+//    if(dspC < 0)
+//        dspC *= -1;
+//    //Check if both starting location and destination are on the board
+//    if(strtR<0 || strtC<0 || desR<0 || desC<0)
+//        vld = false;
+//    else if(strtR >= ROW_MX || strtC >= COL_MX || desR >= ROW_MX || desC >= COL_MX)
+//        vld = false;
+//    //Check if Frontline or Mountains were selected as destination
+//    else if(board[desR][desC].type == 'M' || board[desR][desC].type == 'F')
+//        vld = false;
+//    //Check if location is occupied that it is not players own unit
+//    else if(board[desR][desC].isOcp && board[desR][desC].occUnit->plyrID == id)
+//        vld = false;
+//    //If move is <1,1> then make sure a Camp location is invloved
+//    else if(dspR == 1 && dspC == 1){
+//        if(board[strtR][strtC].type != 'C' && board[desR][desC].type != 'C')
+//            vld = false;
+//    }
+//    else if(dspR > 1 && dspC > 1)    //TODO - place the engineer exception in here later
+//        vld = false;
+//    else if(XOR(dspR > 1, dspC > 1) && board[strtR][strtC].isRR){
+//        if(!board[desR][desC].isRR)
+//            vld = false;
+//        else if(dspR > 1)
+//            vld = isRRopn(board, strtR, strtC, desR, 'r');
+//        else
+//            vld = isRRopn(board, strtR, strtC, desC, 'c');
+//    }
+//    return vld;
+//}
+//bool isRRopn(Location **board, int initR, int initC, int dest, char drctn)
+//{
+//    bool rrOpn = true;
+//    if(drctn == 'c'){
+//        if(dest - initC > 0){
+//            for(int c=initC+1; c < dest && rrOpn; c++)
+//                if(board[initR][c].isOcp || !board[initR][c].isRR)
+//                    rrOpn = false;
+//        }
+//        else if(dest - initC < 0)
+//            for(int c=initC-1; c > dest && rrOpn; c--)
+//                if(board[initR][c].isOcp || !board[initR][c].isRR)
+//                    rrOpn = false;
+//    }
+//    else if(drctn == 'r'){
+//        if(dest - initR > 0){
+//            for(int r=initR+1; r < dest && rrOpn; r++)
+//                if(board[r][initC].isOcp || !board[r][initC].isRR)
+//                    rrOpn = false;
+//        }
+//        else if(dest - initR < 0)
+//            for(int r=initR-1; r > dest && rrOpn; r--)
+//                if(board[r][initC].isOcp || !board[r][initC].isRR)
+//                    rrOpn = false;
+//    }
+//    return rrOpn;
+//}
+//void move(Location **board, int pTurn){
+//    bool vldMve;
+//    int strtR, strtC;   //(R,C) of starting position
+//    int desR, desC;     //(R,C) of destination position
+//    Unit *unit;
+//    //Loop until valid move is entered
+//    do{
+//        //Prompt player for movement
+//        cout << "Enter the location of the piece you want to move. Use R,C format: ";
+//        cin >> strtR; cin.ignore(1000,','); 
+//        cin >> strtC; cin.ignore(1000,'\n');
+//        cout << "Enter your destination in R,C format: ";
+//        cin >> desR; cin.ignore(1000,','); 
+//        cin >> desC; cin.ignore(1000,'\n');
+//        //Validate movement
+//        vldMve = mveVld(board, strtR, strtC, desR, desC, pTurn);
+//        if(!vldMve)
+//            cout << "Invalid Move\n";
+//        if(board[strtR][strtC].isOcp){
+//            unit = board[strtR][strtC].occUnit;
+//            if(unit->plyrID != pTurn){
+//            cout << "That is not your piece!\n";
+//            vldMve = false;
+//            }
+//            else if(unit->name == "Flag" || unit->name == "Landmines"){
+//                cout << unit->name << " cannot ever be moved.\n";
+//                vldMve = false;
+//            }
+//        }
+//        else{
+//            cout << "No piece available to move from that location.\n";
+//            vldMve = false;
+//        }
+//    }while(!vldMve);
+//    //Now make the move, including combat
+//    if(board[desR][desC].isOcp)
+//        combat(unit, board[desR][desC]);
+//    else{
+//        unOcpy(&board[strtR][strtC]);
+//        ocpy(&board[desR][desC], unit);
+//    } 
+//}
+//string frmtCse(string s){
+//    s[0] = toupper(s[0]);
+//    for(int i=1; i<s.length(); i++)
+//        s[i] = tolower(s[i]);
+//
+//    //Proper capitalization for Field Marshal
+//    if(s.length() == 13)
+//        s[6] = toupper(s[6]);
+//    
+//    return s;
+//}
+//int fndMtch(Unit pcs[], string s){
+//    for(int i=0; i < N_PCS; i++){
+//        if(pcs[i].inPlay == false)
+//            if(pcs[i].name == s)
+//                return i; //return value as soon as a match is found
+//    }
+//    return -1;  //Flag for no match found
+//}
+//void whoNtSt(Unit pcs[]){
+//    int perLne = 0;
+//    for(int i=0; i < N_PCS; i++){
+//        if(pcs[i].inPlay == false){
+//            cout << setw(14) << pcs[i].name;
+//            perLne++;
+//            if(0 == (perLne % 3))
+//                cout << endl;
+//        }
+//    }
+//    cout << endl;
+//}
+//void ptBrdLoc(Location **brd){
+//    for(int r=0; r < ROW_MX; r++){
+//        for(int c=0; c < COL_MX; c++)
+//            cout << brd[r][c].type << " ";
+//        cout << endl;
+//        for(int c=0; c < COL_MX; c++)
+//            cout << brd[r][c].isOcp << " ";
+//        cout << endl;
+//        for(int c=0; c < COL_MX; c++)
+//            cout << brd[r][c].isRR << " ";
+//        cout << endl;
+//        for(int c=0; c < COL_MX; c++)
+//            cout << brd[r][c].occUnit << " ";
+//        cout << endl;
+//        for(int i=0; i < RM_CL; i++){
+//            for(int c=0; c < COL_MX; c++)
+//                for(int j=0; j < RM_RW; j++){
+//                    cout << brd[r][c].dsply1[i][j];
+//                    if(j==12)
+//                        cout << "     ";
+//                }
+//            cout << endl;
+//        }
+//    }
+//}
+//void ptPlyrs(Unit p1[], Unit p2[]){
+//    for(int i=0; i < N_PCS; i++){
+//        cout << "P1: " << p1[i].priority << " " << p1[i].inPlay << " " << p1[i].name << endl;
+//        cout << "P2: " << p2[i].priority << " " << p2[i].inPlay << " " << p2[i].name << endl;
+//    }
+//}
